Add 1.12 support (Fixes #20)
