+++ src/com/flobi/floauction/AuctionBid.java
