+++ src/com/flobi/floauction/AuctionLot.java
