+++ src/com/flobi/floauction/AuctionMessageManager.java
