+++ src/com/flobi/floauction/AuctionParticipant.java
