+++ src/com/flobi/floauction/AuctionScope.java
