+++ src/com/flobi/floauction/utilities/Functions.java
